Use a new class called CenteredOriginAxis for representing the two axes (x, y).
This class creates what is essentially a sequence with the specified
number of elements and where the first and last elements are the same
distance from zero but have different signs. The sequence may or may
not include zero depending on whether it has an even or odd number
of elements (odd will have zero, even will not).