Put the symbols for pieces that are used in the message syntax in the Piece objects.
This kind of sucks because the symbols are already used as the keys in the pieces dictionary.
However, I think this is ok because the keys don't really mean anything - they are
only used for quickly accessing a particular piece. They could be anything - I just
happen to have chosen to use the values from the message syntax.