Add functions for doing the initial board population.
Finally, we take advantage of the "placement step" and initial_placement
information for the pieces and the initial_placement value for the spaces
(i.e. whether a piece can be initially placed there) so that we can create
a valid initial board setup for our own pieces (which we do in _do_initial_placement)
and so that we can place placeholder pieces which represent enemy pieces.