Add a scheme by which the events that have happened to a piece can be recorded.
There are two types of events: movements and attacks. Movements are where
the piece itself moves - it can be only a move or it can be attacking some
other piece. The attack events occur when another piece attacks this piece.
These two types of events are tracked both together and separately. The
initial placement is recorded with a special movement event with a start
position of None and a turn value of 0. This way, the initial placement
can be determined by looking at the end value of the first movement event.