Rework the relationship between Movement and BoardPiece.
BoardPiece no longer adds an implicit initial placement event to itself -
this will be added explicitly now. All methods now behave sanely before
the piece has been placed.

Movement is now a real class, not a namedtuple. It is now created by
being passed a LuzhanqiBoard, a BoardPiece, an end coordinate, and
possibly an attack outcome. This is different mainly because the start
position and attacked piece no longer need to be explicitly specified -
they are determined from the BoardPiece and the LuzhanqiBoard.

The new Movement is temporarily lacking turn tracking - this will come
back.