Flush stdout after printing the player to ref messages.
There was a change in behavior in Python 3 - calling the input()
function results in stdout being flushed. I can't find it documented
anywhere, but [0] seems to be the commit to CPython that changed the
behavior. Interestingly, flushing stdout when reading from stdin seems
to be the traditional behavior of C stdio libraries as well, although it
is not required by the standard.[1] glibc has this behavior - I
believe [2] is where it is implemented.

Someone complained (rather obnoxiously) about this (lack of) behavior in
raw_input() back in 02.[3][4]. At the time, those writing Python (including
Guido van Rossum) declined to make the change. Note how the commit I
referenced above was made by the BDFL himself (albeit 5 years later).

Since this change was not documented anywhere, it seems wise not to depend
on it. Therefore, in this commit we add an explicit flush.

[0] http://hg.python.org/cpython/rev/aa73206c9ce6
[1] http://stackoverflow.com/questions/2123528/does-reading-from-stdin-flush-stdout.
[2] https://sourceware.org/git/?p=glibc.git;a=blob;f=libio/fileops.c#l579
[3] https://mail.python.org/pipermail/python-bugs-list/2002-March/010726.html
[4] https://mail.python.org/pipermail/python-list/2002-March/125650.html