Greatly increase the power of the board visualization code.
This is more code and the board that it prints is significantly larger,
but it is far more powerful (allows putting arbitrary markers on
arbitrary positions on the board) and it should be easier to tell
what is in which row/column.

I'm currently using this to verify the railroad code.