Store attacks by keeping track of which piece won the attack.
This is a more useful method than keeping track of the attack outcome
string that we got from the referee. A value of None now means that
neither piece won, i.e. there was a tie. To disambiguate this with
the case when we simply don't know the outcome, a new "theoretical"
field was introduced, which can be True or False.