Rework the way that the move types are generated and kept track of.
The move type strings are now generated in the functions that enumerate the
moves - namely, road_moves and _railroad_moves. In addition, each Movement
can now have multiple types. I think the only time that this can happen in
Luzhanqi is that moves a single position along a railroad can also be
considered road moves.